@@ -1,43 +1,36 @@
 {
-  "name": "chat-backend",
+  "name": "project-template-backend",
   "version": "0.0.1",
-  "description": "Koodi101 chat",
+  "description": "Koodi101 project template",
   "author": "Marko Klemetti <marko.klemetti@gmail.com>",
   "license": "UNLICENSED",
   "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/koodi101/project-template.git"
+  },
   "engines": {
     "node": ">=7"
   },
   "main": "server.js",
   "scripts": {
-    "dev": "LOGGING=true ./node_modules/.bin/nodemon server.js",
-    "start": "node server.js"
+    "dev": "nodemon server.js",
+    "start": "node server.js",
+    "lint": "eslint --ext .js src"
   },
   "dependencies": {
-    "kcors": "^2.2.1",
-    "koa": "^2.2.0",
-    "koa-bodyparser": "^4.2.0",
-    "koa-logger": "^3.1.0",
-    "koa-router": "^7.1.1",
-    "sequelize": "^3.30.4",
-    "sequelize-cli": "^2.7.0",
-    "sqlite3": "^3.1.13",
-    "uglify-js": "^3.2.2"
+    "kcors": "^2.2.2",
+    "koa": "^2.11.0",
+    "koa-bodyparser": "^4.2.1",
+    "koa-logger": "^3.2.1",
+    "koa-router": "^8.0.6",
+    "pg": "^7.18.1",
+    "sequelize": "^5.21.3"
   },
   "devDependencies": {
-    "assert": "^1.4.1",
-    "babel-eslint": "^8.0.3",
-    "eslint": "^4.13.0",
-    "eslint-config-standard": "^11.0.0-beta.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-node": "^5.2.1",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "methods": "^1.1.2",
-    "mocha": "^4.0.1",
-    "nodemon": "^1.11.0",
-    "request": "^2.81.0",
-    "sinon": "^4.1.3",
-    "supertest": "^3.0.0"
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-plugin-import": "^2.20.1",
+    "nodemon": "^2.0.2"
   }
 }
