@@ -1,57 +1,71 @@
-/******** DO NOT DELETE THESE LINES ********/
+/* DO NOT DELETE THESE LINES */
 
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
 
-import './assets/stylesheets/style.css'
+import './assets/stylesheets/style.css';
 
-const baseURL = process.env.ENDPOINT;
+const { BACKEND_PORT } = process.env;
+const baseUrl = window.location.hostname;
+const backendUrl = `http://${baseUrl}:${BACKEND_PORT}`;
 
-/****** ADD YOUR CODE AFTER THIS LINE ******/
+/* ADD YOUR CODE AFTER THIS LINE */
 
 const getGreetingFromBackend = async () => {
   try {
-    const url = `${baseURL}/api/greeting`
-    console.log("Getting greeting from "+url)
+    const url = `${backendUrl}/api/greeting`;
+    console.log(`Getting greeting from ${url}`);
     const response = await fetch(url);
-    return response.json()
+    return response.json();
   } catch (error) {
     console.error(error);
   }
-  return { greeting :"Could not get greeting from backend"};
+  return { greeting: 'Could not get greeting from backend' };
 };
 
 
 const BackendGreeting = (props) => (
-  <div><p>Backend says: {props.greeting}</p></div>
+  <div>
+    <p>
+      Backend says:
+      {' '}
+      {props.greeting}
+    </p>
+  </div>
 );
 
+BackendGreeting.propTypes = {
+  greeting: PropTypes.string,
+};
+
+BackendGreeting.defaultProps = {
+  greeting: '',
+};
 
 class App extends Component {
-
   constructor(props) {
     super(props);
     this.state = {
-      greeting: "",
+      greeting: '',
     };
   }
 
-  async componentWillMount() {
+  async componentDidMount() {
     const response = await getGreetingFromBackend();
-    this.setState({greeting: response.greeting});
+    this.setState({ greeting: response.greeting });
   }
 
   render() {
-
     return (
       <BackendGreeting greeting={this.state.greeting} />
     );
   }
 }
 
-/****** DO NOT DELETE AFTER THIS LINE ******/
+/* DO NOT DELETE AFTER THIS LINE */
 
 ReactDOM.render(
   <App />,
-  document.getElementById('root')
+  document.getElementById('root'),
 );
