@@ -6,9 +6,9 @@ const cors = require('kcors');
 
 const chats = require('./controllers/chats');
 const greeting = require('./controllers/greeting');
-const sensors = require('./controllers/sensors');
+const events = require('./controllers/events');
 
-const app = module.exports = new Koa();
+const app = new Koa();
 
 app.use(logger());
 
@@ -20,10 +20,12 @@ const publicRouter = new Router({ prefix: '/api' });
 publicRouter.post('/chats', chats.create);
 publicRouter.get('/chats', chats.list);
 
-publicRouter.post('/sensors', sensors.create);
-publicRouter.get('/sensors', sensors.list);
+publicRouter.post('/events', events.create);
+publicRouter.get('/events', events.list);
 
 publicRouter.get('/greeting', greeting.greet);
 
 app.use(publicRouter.routes());
 app.use(publicRouter.allowedMethods());
+
+module.exports = app;
