@@ -1,11 +1,11 @@
 
 exports.greet = async (ctx) => {
-  const greetings = ["Hello!", "Guten Tag!", "Bonjour!", "Moro!", "Tere!"];
+  const greetings = ['Hello!', 'Guten Tag!', 'Bonjour!', 'Moro!', 'Tere!'];
 
-  const greeting = greetings[Math.floor(Math.random()*greetings.length)];
+  const greeting = greetings[Math.floor(Math.random() * greetings.length)];
 
-  let response = {
-    greeting: greeting,
+  const response = {
+    greeting,
   };
 
   ctx.body = response;
